+++ internal/telegram/types.go
@@ -7,7 +7,7 @@ type Update struct {
 }
 
 type Message struct {
-	MessageID      int             `json:"message_id"`
+	MessageID      int64           `json:"message_id"`
 	NewChatMembers []User          `json:"new_chat_members"`
 	Chat           Chat            `json:"chat"`
 	From           User            `json:"from"`
@@ -16,14 +16,14 @@ type Message struct {
 }
 
 type User struct {
-	ID        int    `json:"id"`
+	ID        int64  `json:"id"`
 	Username  string `json:"username"`
 	FirstName string `json:"first_name"`
 	LastName  string `json:"last_name"`
 }
 
 type Chat struct {
-	ID   int    `json:"id"`
+	ID   int64  `json:"id"`
 	Type string `json:"type"`
 }
 
@@ -38,19 +38,19 @@ func (e MessageEntity) Text(m *Message) string {
 }
 
 type sendMessagePayload struct {
-	ChatID          int              `json:"chat_id"`
+	ChatID          int64            `json:"chat_id"`
 	Text            string           `json:"text"`
 	ReplyParameters *replyParameters `json:"reply_parameters,omitempty"`
 }
 
 type replyParameters struct {
-	MessageID int
-	ChatID    int
+	MessageID int64
+	ChatID    int64
 }
 
 type SendMessageOption func(*sendMessagePayload)
 
-func WithReplyParameters(messageID int, chatID int) SendMessageOption {
+func WithReplyParameters(messageID int64, chatID int64) SendMessageOption {
 	return func(p *sendMessagePayload) {
 		p.ReplyParameters = &replyParameters{MessageID: messageID, ChatID: chatID}
 	}
